src/tools.py
from langchain_community.tools.tavily_search import TavilySearchResults
from langchain_core.tools import tool

# Tavily search tool for the LeadResearcherAgent
# This tool will use the TAVILY_API_KEY environment variable
tavily_tool = TavilySearchResults(max_results=5)

# Placeholder tool for the EthicalComplianceSentinelAgent
@tool
def content_review_tool(content: str) -> dict:
    """
    Reviews content for compliance and ethical guidelines.
    This is a placeholder and will be implemented later.
    For now, it returns a 'pass' recommendation.
    """
    print("---[Content Review Tool Called]---")
    print(f"Content for review: {content[:100]}...")
    return {"recommendation": "pass", "reasoning": "Placeholder: All content is currently approved."}