pyproject.toml
[project]
name = "nmm-agentic-system"
version = "0.1.0"
description = "Agentic System for New Money Millionaires L.L.C. Operations"
authors = [{ name = "Jules", email = "jules@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain>=0.2.0",
    "langgraph>=0.0.67",
    "langchain-google-genai>=1.0.6",
    "langgraph-checkpoint-sqlite>=0.1.2",
    "tavily-python>=0.3.3",
    "python-dotenv>=1.0.1",
    "ipython" # For interactive development
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"